Add log

+++ aosp_apex_injector.py
@@ -34,9 +34,12 @@ def handle_apex_modules(file_path, aosp_path, lunch_target, target_out_path):
     if apex_emulator_folder and os.path.exists(apex_emulator_folder):
         logging.info(f"Emulator APEX folder found for: {file_path} and {apex_emulator_folder}")
         is_merge_success, log_message = merge_apex_files(apex_emulator_folder, file_path, apex_out_file, lunch_target, aosp_path, target_out_path)
-        os.remove(file_path)
-        shutil.copyfile(apex_out_file, file_path)
-        os.remove(apex_out_file)
+        if os.path.exists(apex_out_file):
+            os.remove(file_path)
+            shutil.copyfile(apex_out_file, file_path)
+            os.remove(apex_out_file)
+        else:
+            shutil.copyfile(org_apex_file, file_path)
         logging.info(f"Merging APEX file complete: {apex_out_file} overwrites {file_path} | {is_merge_success} | {log_message}")
     else:
